Refactor Moq behavior analysis and code actions
Introduce `MockBehaviorDiagnosticAnalyzerBase` for common Moq behavior analysis logic. Refactor `SetExplicitMockBehaviorAnalyzer` to inherit from the new base class. Add `BehaviorType` enum in `Moq.CodeFixes` to represent Moq behaviors. Implement `SetExplicitMockBehaviorCodeAction` for setting explicit mock behavior. Update `SetExplicitMockBehaviorFixer` to use the new code action class. Change `MoqKnownSymbols` members and constructors to `internal` for better encapsulation. These changes enhance code organization, reduce redundancy, and improve maintainability.
