+++ src/Analyzers/MockBehaviorDiagnosticAnalyzerBase.cs
@@ -1,0 +1,81 @@
+﻿using Microsoft.CodeAnalysis.Operations;
+
+namespace Moq.Analyzers;
+
+/// <summary>
+/// Serves as a base class for diagnostic analyzers that analyze mock behavior in Moq.
+/// </summary>
+/// <remarks>
+/// This abstract class provides common functionality for analyzing Moq's MockBehavior, such as registering
+/// compilation start actions and defining the core analysis logic to be implemented by derived classes.
+/// </remarks>
+public abstract class MockBehaviorDiagnosticAnalyzerBase : DiagnosticAnalyzer
+{
+    /// <inheritdoc />
+    public override void Initialize(AnalysisContext context)
+    {
+        context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
+        context.EnableConcurrentExecution();
+
+        context.RegisterCompilationStartAction(RegisterCompilationStartAction);
+    }
+
+    internal abstract void AnalyzeCore(OperationAnalysisContext context, IMethodSymbol target, ImmutableArray<IArgumentOperation> arguments, MoqKnownSymbols knownSymbols);
+
+    private void RegisterCompilationStartAction(CompilationStartAnalysisContext context)
+    {
+        MoqKnownSymbols knownSymbols = new(context.Compilation);
+
+        // Ensure Moq is referenced in the compilation
+        if (!knownSymbols.IsMockReferenced())
+        {
+            return;
+        }
+
+        // Look for the MockBehavior type and provide it to Analyze to avoid looking it up multiple times.
+        if (knownSymbols.MockBehavior is null)
+        {
+            return;
+        }
+
+        context.RegisterOperationAction(context => AnalyzeObjectCreation(context, knownSymbols), OperationKind.ObjectCreation);
+
+        context.RegisterOperationAction(context => AnalyzeInvocation(context, knownSymbols), OperationKind.Invocation);
+    }
+
+    private void AnalyzeObjectCreation(OperationAnalysisContext context, MoqKnownSymbols knownSymbols)
+    {
+        if (context.Operation is not IObjectCreationOperation creation)
+        {
+            return;
+        }
+
+        if (creation.Type is null ||
+            creation.Constructor is null ||
+            !(creation.Type.IsInstanceOf(knownSymbols.Mock1) || creation.Type.IsInstanceOf(knownSymbols.MockRepository)))
+        {
+            // We could expand this check to include any method that accepts a MockBehavior parameter.
+            // Leaving it narrowly scoped for now to avoid false positives and potential performance problems.
+            return;
+        }
+
+        AnalyzeCore(context, creation.Constructor, creation.Arguments, knownSymbols);
+    }
+
+    private void AnalyzeInvocation(OperationAnalysisContext context, MoqKnownSymbols knownSymbols)
+    {
+        if (context.Operation is not IInvocationOperation invocation)
+        {
+            return;
+        }
+
+        if (!invocation.TargetMethod.IsInstanceOf(knownSymbols.MockOf, out IMethodSymbol? match))
+        {
+            // We could expand this check to include any method that accepts a MockBehavior parameter.
+            // Leaving it narrowly scoped for now to avoid false positives and potential performance problems.
+            return;
+        }
+
+        AnalyzeCore(context, match, invocation.Arguments, knownSymbols);
+    }
+}
